Gráficos completos moncal e monpeso

+++ app/src/main/java/com/example/myapplication/MyHelperAcoPeso.java
@@ -27,7 +27,7 @@ public class MyHelperAcoPeso extends SQLiteOpenHelper{
 
     }
 
-    public void insertData(long x, int y){
+    public void insertData(int x, int y){
         SQLiteDatabase db = this.getWritableDatabase();
 
         ContentValues contentValues = new ContentValues();
